@W-14446652@ Change CSP enforcement to an opt-in middleware. (#1528)
* Declare constant in file that uses it.

* Update CHANGELOGs.

* Move CSP middleware to utils.

* Naming things is hard.

* Renaming things is hard.

* Don't touch files that don't need to be touched.

* Re-delete moved code.

* Remove outdated comment.
