+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,6 +1,10 @@
+## v2.1.1 (Nov 7, 2023)
+
+- Use the new opt-in mechanism for enforcing the default security headers required by PWA Kit projects. [#1528](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1528)
+
 ## v2.1.0 (Nov 3, 2023)
 
-- Support Storefront Preview 
+- Support Storefront Preview
   - [#1413](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1413)
   - [#1440](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1440)
 - Show discounted and strikethrough prices when there is a promotion on product detail page [1455](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1455)
